+++ example/test/widget_test.dart
@@ -1,5 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:cloud_firestore_mocks/cloud_firestore_mocks.dart';
+import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
 import 'package:firestore_example/main.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -9,7 +9,7 @@ const MessagesCollection = 'messages';
 void main() {
   testWidgets('shows messages', (WidgetTester tester) async {
     // Populate the mock database.
-    final firestore = MockFirestoreInstance();
+    final firestore = FakeFirebaseFirestore();
     await firestore.collection(MessagesCollection).add({
       'message': 'Hello world!',
       'created_at': FieldValue.serverTimestamp(),
@@ -29,7 +29,7 @@ void main() {
 
   testWidgets('adds messages', (WidgetTester tester) async {
     // Instantiate the mock database.
-    final firestore = MockFirestoreInstance();
+    final firestore = FakeFirebaseFirestore();
 
     // Render the widget.
     await tester.pumpWidget(MaterialApp(